+++ components/dada-ir-sym/src/check/subtype.rs
@@ -1,4 +1,4 @@
-mod chains;
 pub mod is_future;
 pub mod is_numeric;
+mod perms;
 pub mod terms;
